class-7-assignment/pyproject.toml
[project]
name = "class-7-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
]
